refactor: fixed directory structure of domain
