+++ test/unit/domain/consents.test.ts
@@ -28,11 +28,11 @@
  --------------
  ******/
 import { Request } from '@hapi/hapi'
-import { consentDB, scopeDB } from '../../../../src/lib/db'
-import { createAndStoreConsent } from '../../../../src/domain/consents'
+import { consentDB, scopeDB } from '../../../src/lib/db'
+import { createAndStoreConsent } from '../../../src/domain/consents'
 import Logger from '@mojaloop/central-services-logger'
 
-import * as ScopeFunction from '../../../../src/lib/scopes'
+import * as ScopeFunction from '../../../src/lib/scopes'
 
 // Declare Mocks
 const mockInsertConsent = jest.spyOn(consentDB, 'insert')
